+++ plugins/syslog/__init__.py
@@ -1,0 +1,5 @@
+"""syslog plugin package."""
+
+from .plugin import SyslogPlugin
+
+__all__ = ["SyslogPlugin"]